Adjusting gameplay
Game resets to main menu on death. Enemy wait for a moment after an attack so they don't get stuck inside the player. Adjusted health and ammo values. Different enemies have different speeds and damage values
When the boss is defeated the game loads main menu tendrils damage the player on the final map. Sill need to add the proper sprites for enemies.